ZK-52: Moved components/widgets folder;  ServerWidget styling.
